requirements.txt
# =============================================================================
# requirements.txt — Dépendances du projet
# =============================================================================

# API
fastapi
uvicorn
python-multipart
jinja2

# Machine Learning
scikit-learn==1.5.1
xgboost
lightgbm
optuna
joblib
numpy
pandas

# MLflow
mlflow

# Base de données
psycopg2-binary

# Visualisation
matplotlib
seaborn
